fix + move class/function create utils
