+++ mlrun/mlutils/__init__.py
@@ -2,12 +2,13 @@
 
 from .models import (
     get_class_fit,
-    create_class,
-    create_function,
     gen_sklearn_model,
     eval_class_model,
     eval_model_v2,
 )
+
+# for backwards compatibility
+from ..utils.helpers import create_class, create_function
 
 from .plots import (
     gcf_clear,
