+++ mlrun/mlutils/models.py
@@ -48,33 +48,6 @@ def get_class_fit(module_pkg_class: str):
     }
 
 
-def create_class(pkg_class: str):
-    """Create a class from a package.module.class string
-
-    :param pkg_class:  full class location,
-                       e.g. "sklearn.model_selection.GroupKFold"
-    """
-    splits = pkg_class.split(".")
-    clfclass = splits[-1]
-    pkg_module = splits[:-1]
-    class_ = getattr(import_module(".".join(pkg_module)), clfclass)
-    return class_
-
-
-def create_function(pkg_func: list):
-    """Create a function from a package.module.function string
-
-    :param pkg_func:  full function location,
-                      e.g. "sklearn.feature_selection.f_classif"
-    """
-    splits = pkg_func.split(".")
-    pkg_module = ".".join(splits[:-1])
-    cb_fname = splits[-1]
-    pkg_module = __import__(pkg_module, fromlist=[cb_fname])
-    function_ = getattr(pkg_module, cb_fname)
-    return function_
-
-
 def gen_sklearn_model(model_pkg, skparams):
     """generate an sklearn model configuration
