+++ mlrun/runtimes/__init__.py
@@ -27,6 +27,7 @@ from .mpijob import (
 )  # noqa
 from .constants import MPIJobCRDVersions
 from .nuclio import nuclio_init_hook
+
 # for legacy imports (MLModelServer moved from here to /serving)
 from ..serving import MLModelServer  # noqa
 from .sparkjob import SparkRuntime, SparkRuntimeHandler  # noqa
