+++ mlrun/utils/helpers.py
@@ -19,6 +19,7 @@ import sys
 import time
 from datetime import datetime, timezone
 from os import path, environ
+from importlib import import_module
 
 import numpy as np
 import requests
@@ -646,3 +647,30 @@ class RunNotifications:
 
         self._hooks.append(_comment)
         return self
+
+
+def create_class(pkg_class: str):
+    """Create a class from a package.module.class string
+
+    :param pkg_class:  full class location,
+                       e.g. "sklearn.model_selection.GroupKFold"
+    """
+    splits = pkg_class.split(".")
+    clfclass = splits[-1]
+    pkg_module = splits[:-1]
+    class_ = getattr(import_module(".".join(pkg_module)), clfclass)
+    return class_
+
+
+def create_function(pkg_func: list):
+    """Create a function from a package.module.function string
+
+    :param pkg_func:  full function location,
+                      e.g. "sklearn.feature_selection.f_classif"
+    """
+    splits = pkg_func.split(".")
+    pkg_module = ".".join(splits[:-1])
+    cb_fname = splits[-1]
+    pkg_module = __import__(pkg_module, fromlist=[cb_fname])
+    function_ = getattr(pkg_module, cb_fname)
+    return function_
